+++ tests/ValueBuilders/ExpandRangeTest.php
@@ -110,6 +110,9 @@ class ExpandRangeTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::fromTraversable
+     * @covers ::fromString
+     * @covers ::parseString
+     * @covers ::parseRange
      * @dataProvider provideArrayRangesToExpand
      */
     public function testStaticallyExpandArrays($range, $expectedResult)
@@ -130,6 +133,8 @@ class ExpandRangeTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::fromString
+     * @covers ::parseString
+     * @covers ::parseRange
      * @dataProvider provideStringRangesToExpand
      */
     public function testStaticallyExpandStrings($range, $expectedResult)
@@ -150,6 +155,8 @@ class ExpandRangeTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::fromString
+     * @covers ::parseString
+     * @covers ::parseRange
      */
     public function testPreservesRangesInReverseOrder()
     {
@@ -172,6 +179,8 @@ class ExpandRangeTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::fromString
+     * @covers ::parseString
+     * @covers ::parseRange
      * @dataProvider provideIllegalRangeStrings
      *
      * @expectedException GanbaroDigital\TextTools\Exceptions\E4xx_CannotParseRange
@@ -212,6 +221,7 @@ class ExpandRangeTest extends PHPUnit_Framework_TestCase
             [ "0-9", [0, 1, 2, 3, 4, 5, 6, 7, 8, 9] ],
             [ "4-7", [4, 5, 6, 7 ] ],
             [ "8-100", range(8, 100) ],
+            [ "3, 6-8, 10", [3, 6, 7, 8, 10] ],
         ];
     }