+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 * ExpandRange - now supports comma-separated ranges, and single numbers (e.g. "3, 6-8, 10")
+* Filters\FilterForMatchingString - simple non-regex `grep`-like utility class
 
 ## 1.0.0 - Tue Sep 1 2015