+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * ExpandRange - now supports comma-separated ranges, and single numbers (e.g. "3, 6-8, 10")
 * Filters\FilterForMatchingString - simple non-regex `grep`-like utility class
+* Filters\FilterOutMatchingString - simple non-regex `grep -v`-like utility class
 
 ## 1.0.0 - Tue Sep 1 2015